Modify Int and test1.cpp

+++ AD-tests/Int.cpp
@@ -5,10 +5,10 @@ Int::Int (int _val): val(_val) {
     dx = 0;
 }
 
-Int::Int (int _val, int _dx): val(_val), dx(_dx) {
+Int::Int (int _val, long long _dx): val(_val), dx(_dx) {
 }
 
-int Int::get_dx(int arg_id) const {
+long long Int::get_dx(int arg_id) const {
     printf("Catch grad: %d, when val = %d., cmp at arg_id: %d\n", dx, val, arg_id);
     return dx;
 }
